Move model db to internal
